Move derivation of "Send to..." configs for basket and gene list UDs to new util file
